Rename rest.client module to rest.site_client
